+++ langchain-core/src/runnables/base.ts
@@ -1,6 +1,7 @@
 import { z } from "zod";
 import pRetry from "p-retry";
 
+import type { RunnableInterface, RunnableBatchOptions } from "./types.js";
 import {
   CallbackManager,
   CallbackManagerForChainRun,
@@ -17,7 +18,6 @@ import { Serializable } from "../load/serializable.js";
 import {
   IterableReadableStream,
   concat,
-  type IterableReadableStreamInterface,
   atee,
   pipeGeneratorWithSetup,
   AsyncGeneratorWithSetup,
@@ -34,55 +34,9 @@ import { AsyncCaller } from "../utils/async_caller.js";
 import { Run } from "../tracers/base.js";
 import { RootListenersTracer } from "../tracers/root_listener.js";
 import { BaseCallbackHandler } from "../callbacks/base.js";
-import { _RootEventFilter } from "./utils.js";
+import { _RootEventFilter, isRunnableInterface } from "./utils.js";
 import { AsyncLocalStorageProviderSingleton } from "../singletons/index.js";
 import { Graph } from "./graph.js";
-
-/**
- * Base interface implemented by all runnables.
- * Used for cross-compatibility between different versions of LangChain core.
- *
- * Should not change on patch releases.
- */
-export interface RunnableInterface<
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  RunInput = any,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  RunOutput = any,
-  CallOptions extends RunnableConfig = RunnableConfig
-> {
-  lc_serializable: boolean;
-
-  invoke(input: RunInput, options?: Partial<CallOptions>): Promise<RunOutput>;
-
-  batch(
-    inputs: RunInput[],
-    options?: Partial<CallOptions> | Partial<CallOptions>[],
-    batchOptions?: RunnableBatchOptions & { returnExceptions?: false }
-  ): Promise<RunOutput[]>;
-
-  batch(
-    inputs: RunInput[],
-    options?: Partial<CallOptions> | Partial<CallOptions>[],
-    batchOptions?: RunnableBatchOptions & { returnExceptions: true }
-  ): Promise<(RunOutput | Error)[]>;
-
-  batch(
-    inputs: RunInput[],
-    options?: Partial<CallOptions> | Partial<CallOptions>[],
-    batchOptions?: RunnableBatchOptions
-  ): Promise<(RunOutput | Error)[]>;
-
-  stream(
-    input: RunInput,
-    options?: Partial<CallOptions>
-  ): Promise<IterableReadableStreamInterface<RunOutput>>;
-
-  transform(
-    generator: AsyncGenerator<RunInput>,
-    options: Partial<CallOptions>
-  ): AsyncGenerator<RunOutput>;
-}
 
 // TODO: Make `options` just take `RunnableConfig`
 export type RunnableFunc<RunInput, RunOutput> = (
@@ -104,17 +58,6 @@ export type RunnableLike<RunInput = any, RunOutput = any> =
   | RunnableInterface<RunInput, RunOutput>
   | RunnableFunc<RunInput, RunOutput>
   | RunnableMapLike<RunInput, RunOutput>;
-
-export type RunnableBatchOptions = {
-  /** @deprecated Pass in via the standard runnable config object instead */
-  maxConcurrency?: number;
-  returnExceptions?: boolean;
-};
-
-export type RunnableIOSchema = {
-  name?: string;
-  schema: z.ZodType;
-};
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type RunnableRetryFailedAttemptHandler = (error: any) => any;
@@ -906,7 +849,7 @@ export abstract class Runnable<
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   static isRunnable(thing: any): thing is Runnable {
-    return thing ? thing.lc_runnable : false;
+    return isRunnableInterface(thing);
   }
 
   /**
